requirements.txt
streamlit
openai-whisper
python-dotenv
groq
requests
